install tls session in connection impl
We can now install the tls session in the connection implementation. The
lifetimes made this code veeeeery ugly. I do not like it. I am not quite
sure how to make it better. I think that this might just be a place that
we need to decompose things a bit more.