try chained network approach
I want a transaction recording, but I don't think this approach that was
suggested will work. I think that execute transmits a variable number of
packets. Also quite uncertain about how the waker plays into the
dynamics here. I think shifting "pending-transmissions" to only
tranmitting a single packet is going to be a bit of an efficiency hit
with all of the generation that is done at the start of execute.

I added a number of debug statements and they seem to show that execute
is called quite infrequently.